Encode spaces and slashes correctly
The `quote` method's default value for `safe` is '/' so it doesn't encode slashes
into "%2F" which is what the Conjur server expects. Thus, we need to use this
method with no safe characters. We can't use the method `quote_plus` (which encodes
slashes correctly) because it encodes spaces into the character '+' instead of "%20"
as expected by the Conjur server