Add galaxy importer validationt to pipeline
In an effort to publish this integration to Ansible Hub for #56, we attempted
to upload our latest release and failed with the error:

> Invalid collection metadata. Each author in 'authors' list must not be greater
> than 64 characters

This commit fixes the error and adds a check to our pipeline to ensure that
we don't only discover galaxy-importer errors on publish.

This commit also updates the pipeline to always build the release, and not just
on tag or master branch builds, so that we can discover validation errors in
branch builds.